+++ .gitignore
@@ -6,3 +6,4 @@ node_modules
 temp
 .vscode
 dist
+.etmp
