+++ archetype/config/index.js
@@ -1,0 +1,9 @@
+process.env.ENABLE_CSS_MODULE = "true";
+module.exports = {
+  webpack: {
+    cssModuleSupport: true
+  },
+  babel: {
+    enableTypeScript: true
+  }
+};
