+++ package.json
@@ -4,10 +4,10 @@
   "description": "The `Type-18` is a self-defined front-end only project starter which is supported by `React`, `Redux`, `Webpack` and other libraries. The Project starter includes demos for `redux`, `react-redux`, `reselect`, `mobx` and other self-defined classes and components. It has already setup `HMR`, `code-spliting`, `babel-plugin-react-css-modules`. The project also supports tests via `mocha`, `jest` and integration test simulating user activity driven by `selenium` and `chromedriver`",
   "main": "index.js",
   "scripts": {
-    "start": "webpack-dev-server --hot --open --mode development --env.development",
+    "start": "webpack-dev-server --config webpack.config.dev.js --progress",
     "prod": "webpack-dev-server --open --mode production --env.production",
     "dev": "webpack --mode development --env.development",
-    "build": "webpack --mode production --env.production",
+    "build": "webpack --progress",
     "test": "node ./node_modules/mocha/bin/mocha",
     "mocha": "node ./node_modules/mocha/bin/mocha",
     "jest": "node ./node_modules/jest/bin/jest",
@@ -107,6 +107,7 @@
     "big.js": "^5.2.2",
     "bootstrap": ">=4.3.1",
     "font-awesome": "^4.7.0",
+    "funny-react-app-archetype": "0.0.3",
     "global-mercator": "^3.0.3",
     "http-status": "^1.3.1",
     "jquery-ui-dist": "^1.12.1",
