+++ src/client/.babelrc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  extends: "funny-react-app-archetype/config/babel/babelrc-client.js",
+  presets: ["babel-preset-mobx"]
+};
