+++ src/client/images/favicon.ico
