+++ src/client/js/components/People.ts
