+++ src/client/js/components/Window.jsx
@@ -6,9 +6,10 @@ import { tileToBBox, pointToTileFraction, googleToTile } from "global-mercator";
 import { ActionWatcher } from "./utils";
 import { TopComp, TopComp1 } from "../context/ContextShow";
 import { ValidationWrapper } from "./ValidationTemp";
-import custom from "../../css/main.css"; // eslint-disable-line
+import custom from "../../styles/main.css"; // eslint-disable-line
 import bootstrap from "bootstrap/dist/css/bootstrap.min.css"; // eslint-disable-line
 import fontawesome from "font-awesome/css/font-awesome.min.css"; // eslint-disable-line
+import favicon from "../../images/favicon.ico";
 export class Window extends React.Component {
   constructor(props) {
     super(props);
@@ -91,6 +92,9 @@ export class Window extends React.Component {
           toggle header
         </button>
         <button styleName="fontawesome.fa fontawesome.fa-mail-forward" aria-hidden="true" />
+        <div>
+          <img src={favicon}/>
+        </div>
         <div className="div1">
           <input type="text" onChange={this.inputWatcher.watch} />
           {this.state.onInput ? <span style={{ color: "red" }}>inputting ...</span> : ""}
