+++ src/client/js/middlewares/logger.jsx
