+++ src/client/js/react-redux-concise/ReactReduxConcise.jsx
