+++ src/client/js/redux-concise/ReduxConcise.jsx
