+++ src/client/js/reselect-concise/ReselectConcise.jsx
