+++ src/client/js/settings/reducers.jsx
