+++ webpack.config.dev.js
@@ -1,0 +1,2 @@
+process.env.NODE_ENV = "development";
+module.exports = require("funny-react-app-archetype/config/webpack/webpack.config.dev");
